+++ api/src/main/java/com/orange/verify/api/bean/CardType.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
  * 软件充值卡类型
  * t_card_type
  * @author Orange
- * @date 2018/12/22
+ * @date 2018/12/23
  */
 @TableName("t_card_type")
 @KeySequence("SEQ_TEST")
@@ -48,7 +48,7 @@ public class CardType implements Serializable {
      * 卡类值 比如对应的是分填1就是1分钟 以此类推
      * value
      */
-    private Long value;
+    private Integer value;
 
     public String getId() {
         return id;
@@ -106,11 +106,11 @@ public class CardType implements Serializable {
         this.unit = unit;
     }
 
-    public Long getValue() {
+    public Integer getValue() {
         return value;
     }
 
-    public void setValue(Long value) {
+    public void setValue(Integer value) {
         this.value = value;
     }
 }
